Use direct CORs to the TransportAPI.
TransportAPI has enabled direct Cross Origin Requests so the js lib should
use direct JSON requests VS JSONP ones.

All the functions are tested and work like before, and the public interface is
not changed. That's why the proposed version is MINOR -> 0.1.0.

Adding configurable timeout for the requests - some are expected to be slower
than others, for example the journey planning ones.